moved tensor manipulation functions to separate tensor_tools module
